requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
asyncpg==0.29.0
iso8601==1.1.0
pydantic==2.8.0
pydantic_core==2.20.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
tortoise-orm==0.21.3
typing_extensions==4.12.2
